added 'chess' package and its subpackages
