+++ src/main/java/chess/mechanics/Cell.java
@@ -1,3 +1,7 @@
+package chess.mechanics;
+
+import chess.mechanics.pieces.*;
+
 import javax.swing.*;
 import java.awt.*;
