+++ src/main/java/chess/mechanics/Move.java
@@ -1,3 +1,5 @@
+package chess.mechanics;
+
 import java.awt.*;
 import java.io.Serializable;
 import java.util.HashMap;
