+++ src/main/java/chess/network/ConnectionHandler.java
@@ -1,4 +1,10 @@
-import java.io.*;
+package chess.network;
+
+import chess.mechanics.Board;
+
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.net.SocketTimeoutException;
