+++ src/main/java/chess/network/Message.java
@@ -1,10 +1,14 @@
+package chess.network;
+
+import chess.mechanics.*;
+
 import java.io.Serializable;
 
 public class Message implements Serializable {
     public Move move;
     public CheckDetector.State state;
 
-    Message(Move move, CheckDetector.State state) {
+    public Message(Move move, CheckDetector.State state) {
         this.move = move;
         this.state = state;
     }
