+++ src/main/java/chess/utilities/Utility.java
@@ -1,7 +1,9 @@
+package chess.utilities;
+
 import java.io.File;
 
 public class Utility {
-    static String getResourcePath(){
+    public static String getResourcePath(){
         return "src" + File.separator + "main" + File.separator + "resources" + File.separator;
     }
 }
